backend/pyproject.toml
[project]
name = "licence-api"
version = "0.1.0"
description = "License Management System API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "redis>=5.0.0",
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=41.0.0",
    "httpx>=0.26.0",
    "apscheduler>=3.10.0",
    "python-multipart>=0.0.6",
    "bcrypt>=4.0.0",
    "slowapi>=0.1.9",
    "openpyxl>=3.1.0",
    "qrcode[pil]>=7.4.0",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/licence_api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = [
    # SQLAlchemy requires == True/False for column comparisons
    "E712",
    # Forward references in ORM models are valid
    "F821",
    # Module imports at bottom for circular import handling
    "E402",
    # str + Enum inheritance is valid for serialization
    "UP042",
    # Generic[T] syntax is clear and widely used
    "UP046",
    # Alias imports like `UUID as PyUUID` are intentional
    "N811",
]

[tool.ruff.lint.per-file-ignores]
# HTML email templates have long inline-style lines
"src/licence_api/services/email_service.py" = ["E501"]
# ORM models have circular imports resolved at bottom
"src/licence_api/models/orm/*.py" = ["E402", "F401", "F821"]

[tool.mypy]
python_version = "3.12"
strict = true
